requirements.txt
numpy==2.3.2
opencv_python==4.12.0.88
Pillow==11.3.0
torch==2.3.0+cu121
torchvision==0.18.0+cu121
